change decimal aot float

+++ db/migrate/20240105164534_create_carts.rb
@@ -1,7 +1,7 @@
 class CreateCarts < ActiveRecord::Migration[7.1]
   def change
     create_table :carts do |t|
-      t.decimal :total_price, default: 0.0
+      t.float :total_price, default: 0.0
       t.references :user, foreign_key: true
 
       t.timestamps
